Preserve miniapp stream fps
