+++ mobile/modules/island/src/runtime/config.ts
@@ -314,8 +314,7 @@ export interface StreamVideoConfig {
   width?: number
   height?: number
   bitrate?: number
-  /** Request input uses frameRate; resolved stream status reports video.fps. */
-  frameRate?: number
+  fps?: number
 }
 
 export interface StreamAudioConfig {
