+++ mobile/modules/island/src/runtime/streamConfig.ts
@@ -1,0 +1,34 @@
+import type {StreamAudioConfig, StreamVideoConfig} from "./config"
+
+const isRecord = (value: unknown): value is Record<string, unknown> =>
+  typeof value === "object" && value !== null && !Array.isArray(value)
+
+const finiteNumber = (value: unknown): number | undefined =>
+  typeof value === "number" && Number.isFinite(value) ? value : undefined
+
+export const normalizeStreamVideoConfig = (value: unknown): StreamVideoConfig | undefined => {
+  if (!isRecord(value)) return undefined
+  const config: StreamVideoConfig = {}
+  const width = finiteNumber(value.width)
+  const height = finiteNumber(value.height)
+  const bitrate = finiteNumber(value.bitrate)
+  const frameRate = finiteNumber(value.frameRate)
+  const fps = frameRate ?? finiteNumber(value.fps)
+  if (width !== undefined) config.width = width
+  if (height !== undefined) config.height = height
+  if (bitrate !== undefined) config.bitrate = bitrate
+  if (fps !== undefined) config.fps = fps
+  return Object.keys(config).length > 0 ? config : undefined
+}
+
+export const normalizeStreamAudioConfig = (value: unknown): StreamAudioConfig | undefined => {
+  if (!isRecord(value)) return undefined
+  const config: StreamAudioConfig = {}
+  const bitrate = finiteNumber(value.bitrate)
+  const sampleRate = finiteNumber(value.sampleRate)
+  if (bitrate !== undefined) config.bitrate = bitrate
+  if (sampleRate !== undefined) config.sampleRate = sampleRate
+  if (typeof value.echoCancellation === "boolean") config.echoCancellation = value.echoCancellation
+  if (typeof value.noiseSuppression === "boolean") config.noiseSuppression = value.noiseSuppression
+  return Object.keys(config).length > 0 ? config : undefined
+}
